Create Double Slabs default config.
Note: This is to work around a critical bug in the mod that causes lag and crashes when double slabs are used. This is hopefully temporary, and should be reverted once the bug is fixed in the mod.

A side effect of this is that several features of the mod are disabled, but handling it this way will prevent existing worlds from breaking.